requirements.txt
flask==3.0.3
flask-cors==4.0.0
tflite-runtime==2.14.0
numpy==1.26.4
pillow==10.4.0
gunicorn==22.0.0